LayoutTests/fast/scrolling/playstation/scrollIntoView-on-option-expected.txt


PASS scroller.scrollTop is within 10 of 830
PASS listbox.scrollTop is within 2 of 62
PASS successfullyParsed is true

TEST COMPLETE
